feat: random service and refactor of multiple files
- cleaned checklist file
- removed handlers tracing logs
- removed duplicated pi service code
