+++ Cargo.toml
@@ -12,6 +12,7 @@ csv = "1.3.1"
 hickory-proto = "0.25.2"
 hickory-server = "0.25.2"
 once_cell = "1.21.3"
+rand = "0.9.2"
 regex = "1.11.1"
 serde = "1.0.221"
 serde_json = "1.0.145"
