+++ src/services/pi/mod.rs
@@ -23,75 +23,19 @@ impl PiService {
     pub fn new() -> Self {
         Self
     }
-
-    /// Handles Pi constant queries, returning Pi in different formats.
-    ///
-    /// This function provides access to the mathematical constant Pi (π) through
-    /// DNS queries. It supports multiple record types, returning Pi as text,
-    /// IPv4 address, or IPv6 address depending on the query type.
-    ///
-    /// This is a "mathematical constant" service - it provides easy access to Pi
-    /// through DNS, which can be useful for educational purposes, mathematical
-    /// calculations, or as a demonstration of DNS-based data retrieval.
-    ///
-    /// ## Arguments
-    /// * `request` - The DNS request
-    /// * `query_name` - The DNS name being queried
-    /// * `query_type` - The type of DNS record requested (TXT, A, or AAAA)
-    ///
-    /// ## Returns
-    /// * `Some(Record)` - DNS record containing Pi in the requested format
-    /// * `None` - If the query type is not supported
-    pub async fn handle_pi_query(
-        &self,
-        _request: &Request,
-        query_name: &Name,
-        query_type: RecordType,
-    ) -> Option<Record> {
-        match query_type {
-            // Return Pi as text
-            RecordType::TXT => Some(Record::from_rdata(
-                query_name.clone(),
-                PI_TTL,
-                RData::TXT(rdata::TXT::new(vec![
-                    "3.141592653589793238462643383279502884197169".to_string(),
-                ])),
-            )),
-
-            RecordType::A => {
-                // Return Pi as IPv4: 3.141.59.27
-                let pi_ip = std::net::Ipv4Addr::new(3, 141, 59, 27);
-                Some(Record::from_rdata(
-                    query_name.clone(),
-                    PI_TTL,
-                    RData::A(pi_ip.into()),
-                ))
-            }
-
-            RecordType::AAAA => {
-                // Return Pi as IPv6: 3141:5926:5358:9793:2384:6264:3383:2795
-                if let Ok(pi_ipv6) = "3141:5926:5358:9793:2384:6264:3383:2795".parse() {
-                    Some(Record::from_rdata(
-                        query_name.clone(),
-                        PI_TTL,
-                        RData::AAAA(pi_ipv6),
-                    ))
-                } else {
-                    None
-                }
-            }
-            _ => None,
-        }
-    }
 }
 
 #[async_trait]
 impl Service for PiService {
     /// Handles Pi constant queries, returning Pi in different formats.
     ///
-    /// This unified method provides access to the mathematical constant Pi (π) through
+    /// This method provides access to the mathematical constant Pi (π) through
     /// DNS queries. It supports multiple record types, returning Pi as text,
     /// IPv4 address, or IPv6 address depending on the query type.
+    ///
+    /// This is a "mathematical constant" service - it provides easy access to Pi
+    /// through DNS, which can be useful for educational purposes, mathematical
+    /// calculations, or as a demonstration of DNS-based data retrieval.
     ///
     /// ## Arguments
     /// * `request` - The DNS request
